move gravsearch generation methods to its own service
