+++ libs/vre/advanced-search/src/lib/data-access/advanced-search-store/advanced-search-store.service.spec.ts
@@ -4,17 +4,20 @@ import { AdvancedSearchStoreService } from './advanced-search-store.service';
 import { Inject, Injectable, InjectionToken } from '@angular/core';
 import { KnoraApiConnection } from '@dasch-swiss/dsp-js';
 import { AdvancedSearchService } from '../advanced-search-service/advanced-search.service';
+import { GravsearchService } from '../gravsearch-service/gravsearch.service';
 
 export const DspApiConnectionToken = new InjectionToken<KnoraApiConnection>('DspApiConnectionToken');
 
 @Injectable()
-export class MockKnoraApiConnection {
-}
+export class MockKnoraApiConnection {}
 
 @Injectable()
 export class MockAdvancedSearchService {
     constructor(@Inject(DspApiConnectionToken)private _knoraApiConnection: MockKnoraApiConnection) {}
 }
+
+@Injectable()
+export class MockGravsearchService {}
 
 describe('AdvancedSearchStoreService', () => {
     let service: AdvancedSearchStoreService;
@@ -30,6 +33,10 @@ describe('AdvancedSearchStoreService', () => {
                 {
                     provide: AdvancedSearchService,
                     useClass: MockAdvancedSearchService,
+                },
+                {
+                    provide: GravsearchService,
+                    useClass: MockGravsearchService,
                 },
                 AdvancedSearchStoreService
             ],
