+++ libs/vre/advanced-search/src/lib/data-access/gravsearch-service/gravsearch.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { GravsearchService } from './gravsearch.service';
+
+describe('GravsearchService', () => {
+  let service: GravsearchService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(GravsearchService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
